refactor(learning): decompose executor.js into focused modules
SD-LEO-REFACTOR-LEARN-001 (child of SD-LEO-REFACTOR-LARGE-FILES-001)

Extract 1,244 LOC executor.js into 6 focused modules:
- sd-builders.js (178 LOC) - SD field builder functions
- classification.js (126 LOC) - Complexity classification logic
- sd-creation.js (291 LOC) - SD creation workflow
- improvement-appliers.js (406 LOC) - Apply improvement handlers
- decision-management.js (222 LOC) - Decision records and rollback
- executor.js (88 LOC) - Main entry point with re-exports

All modules now under 500 LOC target.
Maintains full backward compatibility via re-exports.
